priceSummaryFunc
