+++ src/components/pages/confirmation.tsx
@@ -1,7 +1,7 @@
 
 import { CSSProperties, FC, useContext } from "react"
 import * as React from 'react';
-import { spaceBetween, summaryCardContainer, priceSummaryFunc } from "../payment/summaryCard";
+import { spaceBetween, summaryCardContainer } from "../payment/summaryCard";
 import { cartContext } from "../context/cartProvider";
 import { colors } from "../../data/color";
 import InvoiceInfoProvider, { invoiceContext } from "../context/invoiceProvider";
@@ -15,6 +15,7 @@ import ReactDOM from "react-dom";
 import { PropsWithChildren, useEffect } from "react";
 import { paymentContext } from "../context/checkOutProvider";
 import RenderPaymentOptions from "../payment/paymentOptions";
+import { priceSummaryFunc } from "../payment/priceLogic";
 
 interface Props {}
